+++ HotSource/src/main/java/hotsource/model/asset/MybatisAssetDAO.java
@@ -20,11 +20,11 @@ public class MybatisAssetDAO implements AssetDAO{
 	}
 
 	@Override
-	public Asset select(int asset_id) {
+	public Asset select(long asset_id) {
 		return sqlSessionTemplate.selectOne("Asset.select", asset_id);
 	}
 	
-	public List selectBySellerId(int seller_id) {
+	public List selectBySellerId(long seller_id) {
 		return sqlSessionTemplate.selectList("Asset.selectBySellerId", seller_id);
 	}
 
@@ -39,7 +39,7 @@ public class MybatisAssetDAO implements AssetDAO{
 	}
 
 	@Override
-	public void delete(int asset_id) {
+	public void delete(long asset_id) {
 		
 	}
