+++ HotSource/src/main/java/hotsource/model/asset_keyword_mapping/AssetKeywordMappingDAO.java
@@ -3,5 +3,5 @@ package hotsource.model.asset_keyword_mapping;
 import java.util.List;
 
 public interface AssetKeywordMappingDAO {
-	public List selectByAssetId(int asset_id);
+	public List selectByAssetId(long asset_id);
 }
