+++ HotSource/src/main/java/hotsource/model/notice_comment/MybatisNoticeCommentDAO.java
@@ -28,7 +28,7 @@ public class MybatisNoticeCommentDAO implements NoticeCommentDAO{
 	}
 
 	@Override
-	public NoticeComment select(int notice_comment_id) {
+	public NoticeComment select(long notice_comment_id) {
 		return sqlSessionTemplate.selectOne("NoticeComment.select", notice_comment_id);
 	}
 
@@ -43,7 +43,7 @@ public class MybatisNoticeCommentDAO implements NoticeCommentDAO{
 	}
 
 	@Override
-	public void delete(int notice_comment_id) {
+	public void delete(long notice_comment_id) {
 		
 	}
 }
