+++ HotSource/src/main/java/hotsource/model/notice_comment/NoticeCommentDAO.java
@@ -7,8 +7,8 @@ import hotsource.domain.NoticeComment;
 
 public interface NoticeCommentDAO {	
 	public List selectAll();
-	public NoticeComment select(int notice_comment_id);
+	public NoticeComment select(long notice_comment_id);
 	public void regist(NoticeComment noticeComment);
 	public void update(NoticeComment noticeComment);
-	public void delete(int notice_comment_id);
+	public void delete(long notice_comment_id);
 }
