+++ HotSource/src/main/java/hotsource/model/review_like/ReviewLikeDAO.java
@@ -3,5 +3,5 @@ package hotsource.model.review_like;
 import java.util.List;
 
 public interface ReviewLikeDAO {
-	public List selectByReviewId(int review_id);
+	public List selectByReviewId(long review_id);
 }
