+++ HotSource/src/main/java/hotsource/model/seller/SellerDAO.java
@@ -6,6 +6,6 @@ import hotsource.domain.Seller;
 
 public interface SellerDAO {
 	public List selectAll();
-	public Seller select(int seller_id);
-	public Seller selectByUserId(int user_id);
+	public Seller select(long seller_id);
+	public Seller selectByUserId(long user_id);
 }
